Added UI
There are now Main Menu, How to Play, and Advanced Movement scenes.
The menus are all fully functional.
Generated lighting for Level 1 and Player Movement Test because I broke it editor-side in the process (don't worry it changes nothing in the final build trust me I have seen it before).